backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
supabase>=2.5.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
structlog>=24.1.0

# AI — google-genai adalah SDK resmi terbaru (bukan google-generativeai yang deprecated)
google-genai>=1.0.0

# Azure OCR (optional, stub mode jika tidak diinstall)
azure-ai-documentintelligence>=1.0.0

# Document forensics
Pillow>=10.0.0
imagehash>=4.3.0
piexif>=1.1.3

# Geo
haversine>=2.8.0
httpx>=0.26.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
